requirements.txt
astroid==1.6.6
awscli==1.16.169
bcrypt==3.1.7
botocore==1.12.159
cffi==1.12.3
colorama==0.3.9
Django==2.0.2
dnspython==1.16.0
docutils==0.14
dominate==2.3.5
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-PyMongo==2.3.0
Flask-WTF==0.14.2
ikp3db==1.1.4
isort==4.3.20
itsdangerous==1.1.0
jedi==0.11.1
jmespath==0.9.4
lazy-object-proxy==1.4.1
mccabe==0.6.1
parso==0.1.1
pbr==5.2.1
pyasn1==0.4.5
pycparser==2.19
pylint==1.8.1
pylint-django==0.8.0
pylint-flask==0.5
pylint-plugin-utils==0.5
pymongo==3.8.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==3.13
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
stevedore==1.30.1
urllib3==1.25.3
virtualenv==16.2.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
visitor==0.1.3
Werkzeug==0.15.4
wrapt==1.11.1
WTForms==2.2.1
